+++ mainlib/ride/actuators.cpp
@@ -1,5 +1,6 @@
 
 #include "actuators.h"
+#include "config/bigringsettings.h"
 #include "model/cyclist.h"
 
 #include "ant/antcentraldispatch.h"
@@ -7,7 +8,9 @@ namespace indoorcycling {
 
 Actuators::Actuators(const Cyclist *cyclist, indoorcycling::AntCentralDispatch *antCentralDispatch,
                                     const NamedSensorConfigurationGroup &sensorConfigurationGroup, QObject *parent):
-    QObject(parent), _cyclist(cyclist), _antCentralDispatch(antCentralDispatch), _sensorConfigurationGroup(sensorConfigurationGroup)
+    QObject(parent), _cyclist(cyclist), _antCentralDispatch(antCentralDispatch), _sensorConfigurationGroup(sensorConfigurationGroup),
+    _maximumSlope(BigRingSettings().maximumUphillForSmartTrainer()),
+    _minimumSlope(-BigRingSettings().maximumDownhillForSmartTrainer())
 {
     connect(_antCentralDispatch, &AntCentralDispatch::sensorFound, this, &Actuators::setSensorFound);
 }
@@ -24,11 +27,12 @@ void Actuators::setSensorFound(AntSensorType channelType, int)
     }
 }
 
-void Actuators::setSlope(qreal slopeInPercent)
+void Actuators::setSlope(const qreal slopeInPercent)
 {
-    if (!qFuzzyCompare(slopeInPercent, _currentSlope)) {
-        _currentSlope = slopeInPercent;
-        _antCentralDispatch->setSlope(slopeInPercent);
+    const qreal cappedSlope = qBound(_minimumSlope, slopeInPercent, _maximumSlope);
+    if (!qFuzzyCompare(cappedSlope, _currentSlope)) {
+        _currentSlope = cappedSlope;
+        _antCentralDispatch->setSlope(cappedSlope);
     }
 }