+++ mainlib/ride/sensors.cpp
@@ -20,7 +20,7 @@
 #include "sensors.h"
 
 #include "ant/antcentraldispatch.h"
-
+#include "config/bigringsettings.h"
 #include <QtCore/QDebug>
 #include <QtCore/QtMath>
 #include <QtCore/QVariant>
@@ -36,7 +36,8 @@ Sensors::Sensors(AntCentralDispatch* antCentralDispatch,
                  const NamedSensorConfigurationGroup &sensorConfigurationGroup,
                  QObject *parent) :
     QObject(parent), _antCentralDispatch(antCentralDispatch),
-    _sensorConfigurationGroup(sensorConfigurationGroup)
+    _sensorConfigurationGroup(sensorConfigurationGroup),
+    _powerForElevationAdjustment(0.01 * BigRingSettings().powerForElevationCorrection())
 {
     if (_sensorConfigurationGroup.simulationSetting() == SimulationSetting::FIXED_POWER ||
         _sensorConfigurationGroup.simulationSetting() == SimulationSetting::VIRTUAL_POWER) {
@@ -121,11 +122,11 @@ void Sensors::sendPowerUpdate()
     if (_sensorConfigurationGroup.simulationSetting() == SimulationSetting::FIXED_POWER) {
         const int fixedPower = _sensorConfigurationGroup.fixedPower();
         const int withRandomPower = fixedPower + (qrand() % 40) - 20;
-        emit powerWattsMeasured(withRandomPower);
+        adjustAndBroadcastPower(withRandomPower);
     } else if (_sensorConfigurationGroup.simulationSetting() == SimulationSetting::VIRTUAL_POWER) {
         qDebug() << "no speed/power input, setting power to 0W.";
         _powerWatts = 0;
-        emit powerWattsMeasured(_powerWatts);
+        emit adjustAndBroadcastPower(_powerWatts);
     }
 }
 
@@ -179,7 +180,7 @@ void Sensors::handlePower(const QVariant &sensorValue, const AntSensorType senso
 
     if (useValue) {
         _powerWatts = sensorValue.toInt();
-        emit powerWattsMeasured(_powerWatts);
+        adjustAndBroadcastPower(_powerWatts);
     }
 }
 
@@ -190,7 +191,7 @@ void Sensors::handleWheelSpeed(const QVariant &sensorValue)
         float wheelSpeedMps = _wheelSpeedRpm * 2.096 / 60.0;
         emit wheelSpeedMpsMeasured(wheelSpeedMps);
     } else if (_sensorConfigurationGroup.simulationSetting() == SimulationSetting::VIRTUAL_POWER) {
-        emit powerWattsMeasured(calculatePower(_wheelSpeedRpm));
+        adjustAndBroadcastPower(calculatePower(_wheelSpeedRpm));
         _updateTimer->start();
     }
 }
@@ -204,5 +205,10 @@ int Sensors::calculatePower(const float wheelSpeedRpm) const
                 .arg(wheelSpeedMps, 2, 'f')
                 .arg(virtualPower, 2, 'f');
     return static_cast<int>(virtualPower);
+}
+
+void Sensors::adjustAndBroadcastPower(int power)
+{
+    emit powerWattsMeasured(static_cast<int>(power * (1.0 + _powerForElevationAdjustment)));
 }
 }